Yet Another Merge with beecrypt-3.0.0.
CVS patchset: 6869
CVS date: 2003/05/22 17:53:03

+++ beecrypt/blockmode.h
@@ -43,7 +43,7 @@ extern "C" {
  * @retval		0 on success.
  */
 BEECRYPTAPI
-int blockEncryptECB(const blockCipher* bc, blockCipherParam* bp, /*@out@*/ uint32_t* dst, const uint32_t* src, size_t nblocks)
+int blockEncryptECB(const blockCipher* bc, blockCipherParam* bp, /*@out@*/ uint32_t* dst, const uint32_t* src, unsigned int nblocks)
 	/*@modifies bp, dst @*/;
 
 /**
@@ -56,7 +56,7 @@ int blockEncryptECB(const blockCipher* bc, blockCipherParam* bp, /*@out@*/ uint3
  * @retval		0 on success.
  */
 BEECRYPTAPI
-int blockDecryptECB(const blockCipher* bc, blockCipherParam* bp, /*@out@*/ uint32_t* dst, const uint32_t* src, size_t nblocks)
+int blockDecryptECB(const blockCipher* bc, blockCipherParam* bp, /*@out@*/ uint32_t* dst, const uint32_t* src, unsigned int nblocks)
 	/*@modifies bp, dst @*/;
 
 /**
@@ -69,7 +69,7 @@ int blockDecryptECB(const blockCipher* bc, blockCipherParam* bp, /*@out@*/ uint3
  * @retval		0 on success.
  */
 BEECRYPTAPI
-int blockEncryptCBC(const blockCipher* bc, blockCipherParam* bp, /*@out@*/ uint32_t* dst, const uint32_t* src, size_t nblocks)
+int blockEncryptCBC(const blockCipher* bc, blockCipherParam* bp, /*@out@*/ uint32_t* dst, const uint32_t* src, unsigned int nblocks)
 	/*@modifies bp, dst @*/;
 
 /**
@@ -82,7 +82,7 @@ int blockEncryptCBC(const blockCipher* bc, blockCipherParam* bp, /*@out@*/ uint3
  * @retval		0 on success.
  */
 BEECRYPTAPI
-int blockDecryptCBC(const blockCipher* bc, blockCipherParam* bp, /*@out@*/ uint32_t* dst, const uint32_t* src, size_t nblocks)
+int blockDecryptCBC(const blockCipher* bc, blockCipherParam* bp, /*@out@*/ uint32_t* dst, const uint32_t* src, unsigned int nblocks)
 	/*@modifies bp, dst @*/;
 
 #ifdef __cplusplus
